pyproject.toml
[project]
name = "bwrp"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Aabish Malik", email = "aabishmalik3337@gmail.com" }
]
dependencies = [
    "kagglehub>=0.3.12",
    "opencv-python>=4.12.0.88",
    "pillow>=11.3.0",
    "matplotlib>=3.10.5",
    "seaborn>=0.13.2",
    "plotly>=6.2.0",
    "pandas>=2.3.1",
    "scikit-learn>=1.7.1",
    "fastapi>=0.116.1",
    "uvicorn>=0.35.0",
    "streamlit>=1.48.0",
    "python-multipart>=0.0.20",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/bwrp"]
